+++ server/src/topic/pubsub.rs
@@ -1,137 +1,192 @@
-use crate::{sink::FanoutMany, BoxSink};
+use super::config::SharedTopicConfig;
+use crate::BoxSink;
+use bytes::Bytes;
 use futures::{
-    channel::{
-        mpsc::{self, Receiver, Sender},
-        oneshot,
-    },
-    ready,
+    channel::mpsc::{self, Receiver, Sender},
     stream::BoxStream,
-    Future, Sink, Stream,
+    Future, SinkExt, StreamExt,
 };
-use log::error;
-use pin_project_lite::pin_project;
-use selium_protocol::traits::{ShutdownSink, ShutdownStream};
-use selium_std::errors::Result;
-use std::{
-    fmt::Debug,
-    pin::Pin,
-    task::{Context, Poll},
+use selium_log::{
+    data::LogIterator,
+    message::{Message, MessageSlice},
+    MessageLog,
 };
+use selium_protocol::{BatchPayload, Frame, MessagePayload, Offset};
+use selium_std::errors::{Result, SeliumError, TopicError};
+use std::{pin::Pin, sync::Arc, time::Duration};
+use tokio::select;
 use tokio_stream::StreamMap;
+use tokio_util::sync::CancellationToken;
+
+pub type SharedLog = Arc<MessageLog>;
+pub type ReadFut = Pin<Box<dyn Future<Output = Result<MessageSlice>> + Send>>;
+pub type SleepFut = Pin<Box<dyn Future<Output = ()> + Send>>;
 
 const SOCK_CHANNEL_SIZE: usize = 100;
 
-pub type TopicShutdown = oneshot::Receiver<()>;
-
-pub enum Socket<T, E> {
-    Stream(BoxStream<'static, Result<T>>),
-    Sink(BoxSink<T, E>),
+pub enum Socket {
+    Stream(BoxStream<'static, Result<Frame>>),
+    Sink(BoxSink<Frame, SeliumError>, Offset),
 }
 
-pin_project! {
-    #[project = TopicProj]
-    #[must_use = "futures do nothing unless you `.await` or poll them"]
-    pub struct Topic<T, E> {
-        #[pin]
-        stream: StreamMap<usize, BoxStream<'static, Result<T>>>,
-        next_stream_id: usize,
-        #[pin]
-        sink: FanoutMany<usize, BoxSink<T, E>>,
-        next_sink_id: usize,
-        #[pin]
-        handle: Receiver<Socket<T, E>>,
-        buffered_item: Option<T>,
+pub struct Subscriber {
+    offset: u64,
+    log: SharedLog,
+    sink: BoxSink<Frame, SeliumError>,
+    buffered_slice: Option<LogIterator>,
+}
+
+impl Subscriber {
+    pub fn new(offset: u64, log: SharedLog, sink: BoxSink<Frame, SeliumError>) -> Self {
+        Self {
+            offset,
+            log: log.clone(),
+            sink,
+            buffered_slice: None,
+        }
+    }
+
+    async fn read_messages(&mut self) {
+        if let Some(slice) = self.buffered_slice.as_mut() {
+            while let Ok(Some(message)) = slice.next().await {
+                let batch_size = message.headers().batch_size();
+                let records = Bytes::copy_from_slice(message.records());
+
+                let frame = if batch_size > 1 {
+                    Frame::BatchMessage(BatchPayload {
+                        message: records,
+                        size: batch_size,
+                    })
+                } else {
+                    Frame::Message(MessagePayload {
+                        headers: None,
+                        message: records,
+                    })
+                };
+
+                let _ = self.sink.send(frame).await;
+            }
+        }
+    }
+
+    async fn poll_for_messages(&mut self, interval: Duration) -> Result<()> {
+        let slice = self
+            .log
+            .read_slice(self.offset, None)
+            .await
+            .map_err(SeliumError::Log)?;
+
+        self.offset = slice.end_offset();
+        self.buffered_slice = slice.messages();
+
+        if self.buffered_slice.is_some() {
+            self.read_messages().await;
+        } else {
+            tokio::time::sleep(interval).await;
+        }
+
+        Ok(())
     }
 }
 
-impl<T, E> Topic<T, E> {
-    pub fn pair() -> (Self, Sender<Socket<T, E>>) {
+pub struct Subscribers {
+    notify: Receiver<Pin<Box<Subscriber>>>,
+    token: CancellationToken,
+    config: SharedTopicConfig,
+}
+
+impl Subscribers {
+    pub fn new(config: SharedTopicConfig) -> (Sender<Pin<Box<Subscriber>>>, Self) {
+        let (tx, notify) = mpsc::channel(SOCK_CHANNEL_SIZE);
+        let token = CancellationToken::new();
+        let subscribers = Self {
+            notify,
+            token,
+            config,
+        };
+        (tx, subscribers)
+    }
+
+    pub async fn run(&mut self) {
+        while let Some(mut subscriber) = self.notify.next().await {
+            let token = self.token.clone();
+            let polling_interval = self.config.polling_interval;
+
+            tokio::spawn(async move {
+                loop {
+                    select! {
+                        _ = token.cancelled() => {
+                            break;
+                        },
+                        _ = subscriber.poll_for_messages(polling_interval) => {
+                            continue;
+                        }
+                    }
+                }
+            });
+        }
+    }
+}
+
+pub struct Topic {
+    publishers: StreamMap<usize, BoxStream<'static, Result<Frame>>>,
+    next_stream_id: usize,
+    notify: Sender<Pin<Box<Subscriber>>>,
+    handle: Receiver<Socket>,
+    log: SharedLog,
+}
+
+impl Topic {
+    pub fn pair(log: MessageLog, config: SharedTopicConfig) -> (Self, Sender<Socket>) {
+        let log = Arc::new(log);
         let (tx, rx) = mpsc::channel(SOCK_CHANNEL_SIZE);
+        let publishers = StreamMap::new();
+        let (notify, mut subscribers) = Subscribers::new(config);
+        tokio::spawn(async move { subscribers.run().await });
 
         (
             Self {
-                stream: StreamMap::new(),
+                log,
+                publishers,
+                notify,
                 next_stream_id: 0,
-                sink: FanoutMany::new(),
-                next_sink_id: 0,
                 handle: rx,
-                buffered_item: None,
             },
             tx,
         )
     }
-}
 
-impl<T, E> Future for Topic<T, E>
-where
-    E: Debug + Unpin,
-    T: Clone + Unpin,
-{
-    type Output = ();
-
-    fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-        let TopicProj {
-            mut stream,
-            next_stream_id,
-            mut sink,
-            next_sink_id,
-            mut handle,
-            buffered_item,
-        } = self.project();
-
+    pub async fn run(&mut self) -> Result<()> {
         loop {
-            // If we've got an item buffered already, we need to write it to the sink
-            // before we can do anything else.
-            if buffered_item.is_some() {
-                // Unwrapping is safe as the underlying sink is guaranteed not to error
-                ready!(sink.as_mut().poll_ready(cx)).unwrap();
-                sink.as_mut()
-                    .start_send(buffered_item.take().unwrap())
-                    .unwrap();
-            }
-
-            match handle.as_mut().poll_next(cx) {
-                Poll::Ready(Some(sock)) => match sock {
-                    Socket::Stream(st) => {
-                        stream.as_mut().insert(*next_stream_id, st);
-                        *next_stream_id += 1;
-                    }
-                    Socket::Sink(si) => {
-                        sink.as_mut().insert(*next_sink_id, si);
-                        *next_sink_id += 1;
+            tokio::select! {
+                Some((_, Ok(frame))) = self.publishers.next() => {
+                    if let Frame::Message(_) | Frame::BatchMessage(_) = frame {
+                        let batch_size = frame.batch_size().unwrap();
+                        let message = frame.message().unwrap();
+                        let message = Message::batch(message, batch_size, 1);
+                        self.log.write(message).await?;
                     }
                 },
-                // If handle is terminated, the stream is dead
-                Poll::Ready(None) => {
-                    ready!(sink.as_mut().poll_flush(cx)).unwrap();
-                    stream.iter_mut().for_each(|(_, s)| s.shutdown_stream());
-                    sink.iter_mut().for_each(|(_, s)| s.shutdown_sink());
+                Some(socket) = self.handle.next() => match socket {
+                    Socket::Stream(st) => {
+                        self.publishers.insert(self.next_stream_id, st);
+                        self.next_stream_id += 1;
+                    }
+                    Socket::Sink(si, offset) => {
+                        let entries = self.log.number_of_entries().await;
 
-                    return Poll::Ready(());
-                }
-                // If no messages are available and there's no work to do, block this future
-                Poll::Pending if stream.is_empty() && buffered_item.is_none() => {
-                    return Poll::Pending
-                }
-                // Otherwise, move on with running the stream
-                Poll::Pending => (),
-            }
+                        let log_offset = match offset {
+                            Offset::FromBeginning(offset) => offset,
+                            Offset::FromEnd(offset) => entries.checked_sub(offset).unwrap_or(entries)
+                        };
 
-            match stream.as_mut().poll_next(cx) {
-                // Received message from an inner stream
-                Poll::Ready(Some((_, Ok(item)))) => *buffered_item = Some(item),
-                // Encountered an error whilst receiving a message from an inner stream
-                Poll::Ready(Some((_, Err(e)))) => {
-                    error!("Received invalid message from stream: {e:?}")
-                }
-                // All streams have finished
-                // Unwrapping is safe as the underlying sink is guaranteed not to error
-                Poll::Ready(None) => ready!(sink.as_mut().poll_flush(cx)).unwrap(),
-                // No messages are available at this time
-                Poll::Pending => {
-                    // Unwrapping is safe as the underlying sink is guaranteed not to error
-                    ready!(sink.poll_flush(cx)).unwrap();
-                    return Poll::Pending;
+                        let subscriber = Box::pin(Subscriber::new(log_offset, self.log.clone(), si));
+
+                        self.notify
+                            .send(subscriber)
+                            .await
+                            .map_err(TopicError::NotifySubscribers)?;
+                    }
                 }
             }
         }