+++ standard/src/codecs/bincode_codec.rs
@@ -1,25 +1,18 @@
+use std::marker::PhantomData;
+
 use crate::traits::codec::{MessageDecoder, MessageEncoder};
 use anyhow::Result;
 use bytes::{Buf, Bytes, BytesMut};
 use serde::{de::DeserializeOwned, Serialize};
-use std::marker::PhantomData;
 
 /// A basic codec that uses [bincode] to serialize and deserialize
 /// binary message payloads.
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub struct BincodeCodec<Item> {
     _marker: PhantomData<Item>,
 }
 
-impl<Item> Clone for BincodeCodec<Item> {
-    fn clone(&self) -> Self {
-        Self {
-            _marker: PhantomData,
-        }
-    }
-}
-
-impl<T> Default for BincodeCodec<T> {
+impl<Item> Default for BincodeCodec<Item> {
     fn default() -> Self {
         Self {
             _marker: PhantomData,
@@ -33,8 +26,10 @@ impl<T> Default for BincodeCodec<T> {
 /// # Errors
 ///
 /// Returns [Err] if `item` fails to serialize.
-impl<Item: Serialize> MessageEncoder<Item> for BincodeCodec<Item> {
-    fn encode(&self, item: Item) -> Result<Bytes> {
+impl<Item: Serialize + Clone> MessageEncoder for BincodeCodec<Item> {
+    type Item = Item;
+
+    fn encode(&self, item: Self::Item) -> Result<Bytes> {
         Ok(bincode::serialize(&item)?.into())
     }
 }
@@ -45,8 +40,10 @@ impl<Item: Serialize> MessageEncoder<Item> for BincodeCodec<Item> {
 /// # Errors
 ///
 /// Returns [Err] if the [BytesMut](bytes::BytesMut) payload fails to deserialize into `Item`.
-impl<Item: DeserializeOwned> MessageDecoder<Item> for BincodeCodec<Item> {
-    fn decode(&self, buffer: &mut BytesMut) -> Result<Item> {
+impl<Item: DeserializeOwned> MessageDecoder for BincodeCodec<Item> {
+    type Item = Item;
+
+    fn decode(&self, buffer: &mut BytesMut) -> Result<Self::Item> {
         Ok(bincode::deserialize_from(buffer.reader())?)
     }
 }